+++ src/js/modal.js
@@ -5,7 +5,6 @@ function modalHandler (e) {
         var $modalEl = $(modalClass)
         modalToggle($modalEl)
     }
-    // console.log('modalHandler activated')
     return false
 }
 
@@ -13,26 +12,21 @@ function modalClose (e) {
     e.preventDefault()
     var $modalEl = $(this).closest('.modal-container')
     modalDisable($modalEl)
-    // console.log('modalClose activated')
 }
 
 function modalToggle ($modalEl) {
     $modalEl.toggleClass('modal-active')
     $('body').toggleClass('modal-active-body')
     $('.modal-mask').toggleClass('active')
-    // console.log('modalToggle activated')
 }
 
 function modalDisable ($modalEl) {
     $modalEl.removeClass('modal-active')
     $('body').removeClass('modal-active-body')
     $('.modal-mask').removeClass('active')
-    // console.log('modalDisable activated')
 }
 
 $(document).ready(function () {
-    var keypressDown = false
-
     $('.modal-link').bind('click', modalHandler)
     $('.modal-close').bind('click', modalClose)
 
@@ -56,13 +50,9 @@ $(document).ready(function () {
     })
 
     $(document).keydown(function (e) {
-        // Esc key
         var keycode = (e.keyCode ? e.keyCode : e.which)
         if (keycode === 27) {
-            if (keypressDown[27] == null) {
-                modalDisable($('.modal-active'))
-                keypressDown[27] = true
-            }
+            modalDisable($('.modal-active'))
         }
     })
 })